+++ src/main/java/urfu/Main.java
@@ -1,36 +1,38 @@
 package urfu;
 
-import urfu.core.CommandInitializer;
-import urfu.core.ICommand;
+import urfu.core.commands.init.CommandInitializer;
+import urfu.core.commands.init.ICommand;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.HashMap;
 
 public class Main
 {
-    public static void main(String[] args) throws IOException
+    public static void main(String[] args)
     {
         System.out.println("[LOG] Bot started...\n");
+
         HashMap<String, ICommand> COMMANDS = CommandInitializer.getAvailableCommands();
 
         BufferedReader INPUT = new BufferedReader(new InputStreamReader(System.in));
 
-        while (true) {
-            System.out.print("notifier@bot: ");
+        try {
+            while (true) {
+                System.out.print("notifier@bot: ");
 
-            // TODO: Надо убедиться, что пользовательский ввод никак и ничего не сможет сломать
-            String userInput = INPUT.readLine();
+                String userInput = INPUT.readLine();
 
-            // TODO: Обрабатывать лишние пробелы, спецсимволы
-            String[] userInputArgs = userInput.split(" ");
+                // TODO: Обрабатывать лишние пробелы, спецсимволы
+                String[] userInputArgs = userInput.split(" ");
 
-            ICommand command = COMMANDS.get(userInputArgs[0]);
+                ICommand command = COMMANDS.get(userInputArgs[0]);
 
-            if (command != null) {
-                command.execute(userInputArgs);
+                if (command != null)
+                    command.execute(userInputArgs);
             }
+        } catch (Exception e) {
+            System.out.println("[ERR] Got exception while waiting user input: " + e.getMessage());
         }
     }
 }
