+++ src/main/java/urfu/core/commands/VersionCommand.java
@@ -1,13 +1,13 @@
 package urfu.core.commands;
 
-import urfu.core.ICommand;
+import urfu.core.commands.init.ICommand;
 
 public class VersionCommand implements ICommand
 {
     public void execute(String[] args)
     {
         // TODO: Сделать вставку значений из .env
-        System.out.println("v1.0.0\n");
+        System.out.println("v1.0-SNAPSHOT\n");
     }
 
     public String getUsageFormat()
